+++ dflib-echarts/src/main/java/org/dflib/echarts/LineSymbol.java
@@ -1,0 +1,8 @@
+package org.dflib.echarts;
+
+/**
+ * @since 2.0.0
+ */
+public enum LineSymbol {
+    emptyCircle, circle, rect, roundRect, triangle, diamond, pin, arrow, none;
+}